Extract tui mod
